+++ sdk/lib/_internal/js_dev_runtime/patch/io_patch.dart
@@ -427,6 +427,11 @@ class InternetAddress {
       InternetAddress address, String host) {
     throw UnsupportedError("InternetAddress._cloneWithNewHost");
   }
+
+  @patch
+  static InternetAddress tryParse(String address) {
+    throw UnsupportedError("InternetAddress.tryParse");
+  }
 }
 
 @patch