+++ sdk_nnbd/lib/_internal/js_runtime/lib/io_patch.dart
@@ -425,6 +425,11 @@ class InternetAddress {
       InternetAddress address, String host) {
     throw new UnsupportedError("InternetAddress._cloneWithNewHost");
   }
+
+  @patch
+  static InternetAddress? tryParse(String address) {
+    throw UnsupportedError("InternetAddress.tryParse");
+  }
 }
 
 @patch